+++ R/dplyr-count.R
@@ -1,7 +1,8 @@
 #' Count the observations in each group
 #' 
 #' `count()` lets you quickly count the unique values of one or more variables. 
-#' It is evaluated lazily.
+#' It is evaluated lazily. `add_count()` is an equivalent that uses `mutate()`
+#' to add a new column with group-wise counts.
 #' @name count.data_request
 #' @param x An object of class `data_request`, created using [galah_call()]
 #' @param wt currently ignored
@@ -14,10 +15,36 @@ count.data_request <- function(x,
                                wt, 
                                sort, 
                                name){
+  count_switch(x) |>
+    group_by(...) |>
+    update_request_object(count = TRUE)
+}
+
+#' Internal function called by `count.data_request()` and `distinct.data_request()`
+#' @noRd
+#' @keywords Internal
+count_switch <- function(x){ 
   x$type <- switch(x$type, 
                    "occurrences" = "occurrences-count",
+                   "occurrences-count" = "occurrences-count",
                    "species" = "species-count",
+                   "species-count" = "species-count",
                    "media" = cli::cli_abort("type = 'media' is not supported by `count()`"),
                    cli::cli_abort("`count()` only supports `type = 'occurrences' or` `'species'`"))
-  group_by(x, ...)
+  x
+}
+
+#' @rdname count.data_request
+#' @export
+add_count.data_request <- function(x,
+                                  ...,
+                                   wt = NULL,
+                                   sort = FALSE,
+                                   name = NULL){
+  # note: this function effectively is only used by `atlas_species()`/`distinct()`
+  # unclear whether this error message will be evaluated properly at this point
+  if(x$type != "species"){
+    cli::cli_abort("`add_count()` is only supported for `type = 'species'`")
+  }
+  update_request_object(x, count = TRUE)
 }